extracted queries into components
