+++ src/Components/Review/Queries/OrderBy.jsx
@@ -1,0 +1,16 @@
+export default function OrderBy ({setOrderBy}) {
+
+    const handleOrderChange = (event) => {
+        setOrderBy(event.target.value);
+    };
+
+    return (
+        <>
+            Order: {}
+            <select onChange={handleOrderChange}>
+                <option value="DESC">Descending</option>
+                <option value="ASC">Ascending</option>
+            </select><br></br>     
+        </>
+    )
+}
