+++ src/apis.js
@@ -25,6 +25,5 @@ export const getCommentsById = (id) => {
 };
 
 export const postComment = (id, username, body) => {
-    console.log(id, username, body);
     return axios.post(`https://neat-games.herokuapp.com/api/reviews/${id}/comments`, {username, body})
 };
